Runtime Event Drawer Setup With Method Selection
Setup the Runtime Event Drawer to maintain a cached collection of Reorderable list objects that contain the different persistent methods that can be selected. Added the menu functionality to generate a list of the available functions and select the option, applying the default values to the persistent callback. Need to implement the parameter drawing functionality to allow for the modification of supplied data values